Moves SampleHandlerBase::Functional helper class to detail subdir
